Add musicplayer and fix bugs, delete SetupCommand.java
We added a Musicplayer with the commands:

play: plays the sound of a video from youtube,
stop: stops the sound,
ti: info about the track that plays currently,
shuffle: shuffle the playlist

Also, we delete the SetupCommand because it's useless. Please configure your prefix and token manually in the config file.

Issues #4 and #5 are in work.